requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.7
idna==3.3
requests==2.26.0
urllib3==1.26.7
xmltodict==0.12.0
chardet
click
future
Jinja2
MarkupSafe
oauthlib
requests-oauthlib==1.0.0
requests-toolbelt==0.8.0
webexteamssdk
Werkzeug
flask==1.1.1
python-dotenv